modified to work better with attached volumes
Instead of downloading the data during the build step, it downloads the
data by default at run time using CMD.  This can be overridden.  This
is required in order to use an attached volume.  To build it into the container,
you have to build the data directory into the container as well, and any attempt
to attach a different volume to it will hide the downloaded data.